Move more types to the common dir
