+++ packages/react-native-reanimated/src/css/component/AnimatedComponent.tsx
@@ -4,7 +4,7 @@ import { Component } from 'react';
 import type { StyleProp } from 'react-native';
 import { Platform, StyleSheet } from 'react-native';
 
-import type { AnyRecord } from '../../common';
+import type { AnyComponent, AnyRecord, PlainStyle } from '../../common';
 import { IS_JEST, ReanimatedError, SHOULD_BE_USE_WEB } from '../../common';
 import type {
   InternalHostInstance,
@@ -20,7 +20,7 @@ import { getShadowNodeWrapperFromRef } from '../../fabricUtils';
 import { findHostInstance } from '../../platform-specific/findHostInstance';
 import { markNodeAsRemovable, unmarkNodeAsRemovable } from '../native';
 import { CSSManager } from '../platform';
-import type { AnyComponent, CSSStyle, PlainStyle } from '../types';
+import type { CSSStyle } from '../types';
 import { filterNonCSSStyleProps } from './utils';
 
 export type AnimatedComponentProps = Record<string, unknown> & {
