+++ packages/react-native-reanimated/src/css/models/CSSKeyframesRuleBase.ts
@@ -1,10 +1,7 @@
 'use strict';
+import type { PlainStyle } from '../../common';
 import { ANIMATION_NAME_PREFIX } from '../constants';
-import type {
-  CSSAnimationKeyframes,
-  CSSKeyframesRule,
-  PlainStyle,
-} from '../types';
+import type { CSSAnimationKeyframes, CSSKeyframesRule } from '../types';
 
 export default abstract class CSSKeyframesRuleBase<S extends PlainStyle>
   implements CSSKeyframesRule
