+++ packages/react-native-reanimated/src/css/native/style/config.ts
@@ -1,11 +1,11 @@
 'use strict';
+import type { PlainStyle } from '../../../common';
 import {
   IS_ANDROID,
   processBoxShadowNative,
   processColor,
   processTransformOrigin,
 } from '../../../common';
-import type { PlainStyle } from '../../types';
 import {
   processAspectRatio,
   processFontWeight,
