+++ packages/react-native-reanimated/src/css/stylesheet/keyframes.ts
@@ -1,10 +1,7 @@
 'use strict';
+import type { PlainStyle } from '../../common';
 import { CSSKeyframesRuleImpl } from '../platform';
-import type {
-  CSSAnimationKeyframes,
-  CSSKeyframesRule,
-  PlainStyle,
-} from '../types';
+import type { CSSAnimationKeyframes, CSSKeyframesRule } from '../types';
 
 export default function keyframes<S extends PlainStyle>(
   // TODO - think of better types
