+++ packages/react-native-reanimated/src/css/types/animation.ts
@@ -1,6 +1,7 @@
 'use strict';
+import type { PlainStyle } from '../../common';
 import type { CSSTimingFunction } from '../easing';
-import type { PlainStyle, TimeUnit } from './common';
+import type { TimeUnit } from './common';
 import type { AddArrayPropertyType, AddArrayPropertyTypes } from './helpers';
 
 export interface CSSKeyframesRule {
