+++ packages/react-native-reanimated/src/css/types/common.ts
@@ -1,27 +1,5 @@
 'use strict';
-import type { ComponentType } from 'react';
-import type {
-  ImageStyle,
-  TextStyle,
-  TransformsStyle,
-  ViewStyle,
-} from 'react-native';
-
-type DeprecatedProps =
-  | 'transformMatrix'
-  | 'rotation'
-  | 'scaleX'
-  | 'scaleY'
-  | 'translateX'
-  | 'translateY';
-
-export type PlainStyle = Omit<
-  ViewStyle & TextStyle & ImageStyle,
-  DeprecatedProps
->;
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type AnyComponent = ComponentType<any>;
+import type { TransformsStyle } from 'react-native';
 
 export type TimeUnit = `${number}s` | `${number}ms` | number;
